Separated modals into files
